Custom sdk log levels:
Added a new SdkLogger type, that behaves the same way, but is really zerolog under the hood and has support for custom log levels for subloggers.

Added a log level manager to the baseapp, which can be configured with custom module names and a log level.

Changed each sdk module to derive its log level when initializing the module logger - unfortunately it's probably not possible to derive without changing module code